# rotas-openapi.yaml
openapi: 3.0.4
info:
  title: APDIS - Rotas API
  description: |-
    Representação da arquitetura REST da plataforma APDIS, desenvolvida no âmbito da unidade curricular de Projeto de Software Web, Móvel e na Nuvem.
    Neste documento será possível observar as rotas existentes, os campos necessários e respostas esperadas.
  version: 1.0.0
servers:
  - url: https://projeto.matz.pt/api
...
paths:
...
  /members/getAllMembers:
    get:
      tags:
        - members
      summary: Obtém todos os membros exceto o próprio utilizador autenticado.
      description: Retorna uma lista de todos os utilizadores registados na plataforma, excluindo o próprio.
      responses:
        '200':
          description: Lista de membros
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    number:
                      type: string
                    fullname:
                      type: string
                    email:
                      type: string
                    type:
                      type: string
        '400':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Não Autorizado."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro: [detalhes]"
  /members/getAllRegisters:
    get:
      tags:
        - members
      summary: Obtém todos os pedidos de registo de associados.
      description: Retorna uma lista de todos os registos pendentes na tabela de requests.
      responses:
        '200':
          description: Lista de registos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '400':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Não Autorizado."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro: [detalhes]"
  /members/approveRegister:
    post:
      tags:
        - members
      summary: Aprova um pedido de registo.
      description: Aprova o registo de um associado com base no ID e Número fornecidos.
      requestBody:
        description: Dados necessários para aprovar o pedido.
        content:
          application/json:
            schema:
              required:
                - id
                - number
              type: object
              properties:
                id:
                  type: integer
                  example: 5
                number:
                  type: string
                  example: "A105"
        required: true
      responses:
        '200':
          description: Pedido aprovado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilizador Aceite."
        '400':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Não Autorizado."
        '500':
          description: Erro interno ou registo não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro."
  /members/rejectRegister:
    delete:
      tags:
        - members
      summary: Rejeita um pedido de registo.
      description: Remove o pedido de um associado com base no ID e Número fornecidos.
      requestBody:
        description: Dados necessários para rejeitar o pedido.
        content:
          application/json:
            schema:
              required:
                - id
                - number
              type: object
              properties:
                id:
                  type: integer
                  example: 7
                number:
                  type: string
                  example: "A107"
        required: true
      responses:
        '200':
          description: Pedido rejeitado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilizador Rejeitado."
        '400':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Não Autorizado."
        '500':
          description: Erro interno ou registo não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro."
  /members/adminMember:
    post:
      tags:
        - members
      summary: Torna um associado administrador.
      description: Altera o tipo de um associado para "admin".
      requestBody:
        description: Dados do associado a promover.
        content:
          application/json:
            schema:
              required:
                - id
                - number
              type: object
              properties:
                id:
                  type: integer
                  example: 12
                number:
                  type: string
                  example: "A120"
        required: true
      responses:
        '200':
          description: Utilizador promovido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilizador é Agora Administrador."
        '400':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Não Autorizado."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro."
  /members/unAdminMember:
    post:
      tags:
        - members
      summary: Remove permissões de administrador.
      description: Altera o tipo de um associado de "admin" para "user".
      requestBody:
        description: Dados do associado a rebaixar.
        content:
          application/json:
            schema:
              required:
                - id
                - number
              type: object
              properties:
                id:
                  type: integer
                  example: 13
                number:
                  type: string
                  example: "A121"
        required: true
      responses:
        '200':
          description: Utilizador atualizado para "user"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilizador Já Não é Administrador."
        '400':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Não Autorizado."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro."
  /members/removeMember:
    delete:
      tags:
        - members
      summary: Remove um associado da plataforma.
      description: Elimina permanentemente um associado com base no ID e Número.
      requestBody:
        description: Dados do associado a remover.
        content:
          application/json:
            schema:
              required:
                - id
                - number
              type: object
              properties:
                id:
                  type: integer
                  example: 10
                number:
                  type: string
                  example: "A110"
        required: true
      responses:
        '200':
          description: Associado removido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilizador Apagado com Sucesso."
        '400':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Não Autorizado."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro."
  /members/getAllLogs:
    get:
      tags:
        - members
      summary: Obtém todos os logs da plataforma.
      description: Retorna todas as entradas de logs de atividade.
      responses:
        '200':
          description: Lista de logs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '400':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Não Autorizado."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro."
  /members/pedidoAssociado:
    post:
      tags:
        - members
      summary: Submete um novo pedido de associação.
      description: Envia os dados de um potencial associado para aprovação.
      requestBody:
        description: Dados pessoais e de associado.
        content:
          application/json:
            schema:
              required:
                - pessoal
                - associado
              type: object
              properties:
                pessoal:
                  type: object
                associado:
                  type: object
        required: true
      responses:
        '200':
          description: Pedido submetido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Pedido Enviado com Sucesso."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro."